fix: Handle nested class names in annotation processor
* and cleanup doc-snippets so that it can be run and tested
