+++ samples/doc-snippets/pom.xml
@@ -5,7 +5,7 @@
   <parent>
     <groupId>io.akka</groupId>
     <artifactId>akka-javasdk-parent</artifactId>
-    <version>3.4.0-M4</version>
+    <version>3.4.0-M5-17-e0d49a17-dev-SNAPSHOT</version>
   </parent>
 
   <groupId>com.example</groupId>
